Refactor macro’s into trait
