+++ packages/core/src/Base/Traits/HasMacros.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace GetCandy\Base\Traits;
+
+use Illuminate\Support\Traits\Macroable;
+
+trait HasMacros
+{
+    use Macroable {
+        __call as macroCall;
+    }
+
+    /**
+     * Handle dynamic method calls into the model.
+     *
+     * @param  string  $method
+     * @param  array  $parameters
+     * @return mixed
+     */
+    public function __call($method, $parameters)
+    {
+        if (static::hasMacro($method)) {
+            return $this->macroCall($method, $parameters);
+        }
+
+        return parent::__call($method, $parameters);
+    }
+
+    /**
+     * Handle dynamic static method calls into the model.
+     *
+     * @param  string  $method
+     * @param  array  $parameters
+     * @return mixed
+     */
+    public static function __callStatic($method, $parameters)
+    {
+        return parent::__callStatic($method, $parameters);
+    }
+}
