+++ packages/core/src/Models/CustomerGroup.php
@@ -4,6 +4,7 @@ namespace GetCandy\Models;
 
 use GetCandy\Base\BaseModel;
 use GetCandy\Base\Traits\HasDefaultRecord;
+use GetCandy\Base\Traits\HasMacros;
 use GetCandy\Base\Traits\HasMedia;
 use GetCandy\Database\Factories\CustomerGroupFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -13,6 +14,7 @@ class CustomerGroup extends BaseModel
     use HasFactory;
     use HasMedia;
     use HasDefaultRecord;
+    use HasMacros;
 
     /**
      * {@inheritDoc}
