+++ packages/core/src/Models/Price.php
@@ -4,12 +4,14 @@ namespace GetCandy\Models;
 
 use GetCandy\Base\BaseModel;
 use GetCandy\Base\Casts\Price as CastsPrice;
+use GetCandy\Base\Traits\HasMacros;
 use GetCandy\Database\Factories\PriceFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Price extends BaseModel
 {
     use HasFactory;
+    use HasMacros;
 
     /**
      * Return a new factory instance for the model.
