+++ packages/core/src/Models/TaxClass.php
@@ -4,6 +4,7 @@ namespace GetCandy\Models;
 
 use GetCandy\Base\BaseModel;
 use GetCandy\Base\Traits\HasDefaultRecord;
+use GetCandy\Base\Traits\HasMacros;
 use GetCandy\Database\Factories\TaxClassFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
@@ -11,6 +12,7 @@ class TaxClass extends BaseModel
 {
     use HasFactory;
     use HasDefaultRecord;
+    use HasMacros;
 
     /**
      * Return a new factory instance for the model.
