+++ packages/core/src/Models/TaxZoneCountry.php
@@ -3,12 +3,14 @@
 namespace GetCandy\Models;
 
 use GetCandy\Base\BaseModel;
+use GetCandy\Base\Traits\HasMacros;
 use GetCandy\Database\Factories\TaxZoneCountryFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class TaxZoneCountry extends BaseModel
 {
     use HasFactory;
+    use HasMacros;
 
     /**
      * Return a new factory instance for the model.
